refactor navigation folder - using index file and move useMainNavigation to hooks folder
