+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useDisclose } from './useDisclose'
 export { default as useKeyboard } from './useKeyboard'
 export { default as useLayout } from './useLayout'
+export { default as useMainNavigation } from './useMainNavigation'
 export { default as useSinglePress } from './useSinglePress'
