+++ src/hooks/useMainNavigation.ts
@@ -1,0 +1,8 @@
+import { MainParamList } from '@/navigators/MainNavigator'
+import { DrawerNavigationProp } from '@react-navigation/drawer'
+import { useNavigation } from '@react-navigation/native'
+
+const useMainNavigation = () =>
+  useNavigation<DrawerNavigationProp<MainParamList>>()
+
+export default useMainNavigation
