+++ src/navigators/MainNavigator/Header.tsx
@@ -3,10 +3,9 @@ import { Pressable, StatusBar } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 import { H4, HStack, Icon } from '@/components'
+import { useMainNavigation } from '@/hooks'
 import { styleManager } from '@/libs'
 import { useRoute } from '@react-navigation/native'
-
-import { useMainNavigation } from './'
 
 const Header = () => {
   const route = useRoute()
