+++ src/navigators/MainNavigator/index.ts
@@ -1,0 +1,2 @@
+export { default, MainRoutes } from './MainNavigator'
+export type { MainParamList } from './MainNavigator'
